+++ spec/services/import/synchronizers/repositories/gitlab_spec.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal: true
 
-describe Import::Repositories::Gitlab, type: :service do
-  subject(:service_call) { described_class.new.call(repository: repository) }
+describe Import::Synchronizers::Repositories::Gitlab, type: :service do
+  subject(:service_call) { instance.call(repository: repository) }
 
+  let!(:instance) { described_class.new }
   let!(:repository) { create :repository, accessable: false, pull_requests_count: 3 }
   let(:insights_service) { double }
 
@@ -10,9 +11,9 @@ describe Import::Repositories::Gitlab, type: :service do
     create :pull_request, repository: repository
     create :pull_request, repository: repository, pull_closed_at: 1.minute.ago
 
-    allow(Pullmetry::Container.resolve('services.import.pull_requests.gitlab')).to receive(:call)
-    allow(Pullmetry::Container.resolve('services.import.comments.gitlab')).to receive(:call)
-    allow(Pullmetry::Container.resolve('services.import.reviews.gitlab')).to receive(:call)
+    allow(Pullmetry::Container.resolve('services.import.synchronizers.pull_requests.gitlab')).to receive(:call)
+    allow(Pullmetry::Container.resolve('services.import.synchronizers.comments.gitlab')).to receive(:call)
+    allow(Pullmetry::Container.resolve('services.import.synchronizers.reviews.gitlab')).to receive(:call)
     allow(Pullmetry::Container.resolve('dummy')).to receive(:call)
 
     allow(Insights::Generate::RepositoryService).to receive(:new).and_return(insights_service)