+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Modified
 - more forms and services with container
+- import services structure
 
 ## [1.5.1] - 2023-09-26
 ### Added